+++ client/src/components/Footer/style.css
@@ -1,0 +1,8 @@
+.footer {
+  position: fixed;
+  bottom: 0;
+  width: 100%;
+  height: 60px;
+  line-height: 60px;
+  background-color: rgba(206, 153, 4, 0.3);
+}
